Use separate creation and update feeds for article submission. Add jitter to start time.
Adding jitter, as well as submitting new articles every minute rather than every five
minutes, will avoid situations where a massive amount of polling jobs are all added
to the queue at exactly the same time. The scheduler and the pollers can handle these
kinds of stampedes, but the external APIs we query might not like it so much.